# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /pastport/guide:
    post:
      operationId: askPastport
      tags: [pastport]
      summary: Ask the contextual historical guide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskPastportBody"
      responses:
        "200":
          description: Contextual answer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PastportGuideResponse"
        "400":
          description: Invalid question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AskPastportBody:
      type: object
      properties:
        question:
          type: string
        site:
          type: string
        year:
          type: integer
        object:
          type: string
      required:
        - question
        - site
        - year
    PastportGuideResponse:
      type: object
      properties:
        answer:
          type: string
        documentedHistory:
          type: boolean
        reconstructionNote:
          type: string
      required:
        - answer
        - documentedHistory
        - reconstructionNote
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message