# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Ask the contextual historical guide
 */
export const AskPastportBody = zod.object({
  "question": zod.string(),
  "site": zod.string(),
  "year": zod.number().int(),
  "object": zod.string().optional()
})

export const AskPastportResponse = zod.object({
  "answer": zod.string(),
  "documentedHistory": zod.boolean(),
  "reconstructionNote": zod.string()
})