# lib/api-zod/src/generated/types/askPastportBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AskPastportBody {
  question: string;
  site: string;
  year: number;
  object?: string;
}