# lib/api-zod/src/generated/types/pastportGuideResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PastportGuideResponse {
  answer: string;
  documentedHistory: boolean;
  reconstructionNote: string;
}